Handling more number types

+++ ExcelORM/ExcelORMTests/TestTypes.cs
@@ -5,6 +5,6 @@ public record TestTypes
     public string? Text { get; set; }
     public DateTime? Date { get; set; }
     public TimeSpan? TimeSpan { get; set; }
-    public double? Int { get; set; }
+    public int? Int { get; set; }
     public double? Double { get; set; }
 }
